Fixed tests post refactor
